[step4] 자동차 경주(우승자) (#179)
* refactor: setter private

* refactor: kotlin random 사용

* refactor: ++ overloading

* refactor: 생성자 기본값 사용

* refactor: 이름충돌 해소

* refactor: rename

* refactor: introduce backing property

* refactor: use repeat

* refactor: 객체생성 validation 적용

* docs: 요구사항 정리

* test: 쉼표로 구분된 문자를 입력받는다

* feat: make it pass

* test: 이름은 5자를 초과할 수 없다

* feat: make it pass

* test: 이름을 생성할 수 있다

* feat: make it pass

* refactor: 구현클래스 이동

* test: 이름과 위치를 출력한다

* feat: Position 에 이름을 부여한다

NamedPosition 을 도입하고 기존 Position 의 이름은 공백처리한다

* refactor: NamedPosition 을 적용한다

* feat: 이름목록 적용

* test: 우승자 목록을 제공한다

* feat: make it pass

* feat: 실제 객체기반 구현

* refactor: 구현클래스 이동

* test: 우승자를 출력한다

* feat: make it pass

* refactor: 구현클래스 이동

* feat: Main 에 결합

* refactor: 위임

* refactor: reorder

* refactor: 연산자 오버로딩

* refactor: Cars 내부 자료구조 변경

이름을 NamedPosition 으로부터 획득

* refactor: copy 활용

* refactor: reorder

* refactor: use inline class

+++ src/test/kotlin/calculation/MathTest.kt
@@ -27,17 +27,17 @@ class MathTest {
 
     @Test
     fun `예약되지 않은 연산자`() {
-        assertThrows<IllegalArgumentException> { Calculation("4 $ 2").toInt() }
+        assertThrows<IllegalArgumentException> { Calculation("4 $ 2") }
     }
 
     @Test
     fun `계산식이 없는 경우`() {
-        assertThrows<IllegalArgumentException> { Calculation("").toInt() }
+        assertThrows<IllegalArgumentException> { Calculation("") }
     }
 
     @Test
     fun `계산식이 비정상인 경우`() {
-        assertThrows<IllegalArgumentException> { Calculation("1 1").toInt() }
+        assertThrows<IllegalArgumentException> { Calculation("1 1") }
     }
 
     @Test
